requirements.txt
fastapi
uvicorn
python-dotenv
aiofiles
pydantic
httpx
groq
google-generativeai
gunicorn
aiosmtplib
streamlit
pandas
google-play-scraper
